% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Dummy function writing "hello world"}
\usage{
hello()
}
\arguments{
\item{retried}{the number of previously done before the exponential backoff sleep}
}
\description{
Dummy function writing "hello world"
}
