\name{Homework1}
\alias{homework1}
\docType{package}
\title{
Fast linear model fit and  multivariate normal (log)density calculation.
}
\description{
This package gives two functions, fastlm() and dmvnorm(), that fast fit a linear regression model and fast calculate the multivariate normal density or log density, respectively, making use of cholesky decomposition of a positive-definite symmetric matrix. 
}
\details{
\tabular{ll}{
Package: \tab homework\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-11-06\cr
}

}
\author{
Yuxin(Daisy) Zhu

Maintainer:Yuxin(Daisy) Zhu <daisy.zhu.yx@gmail.com>

}
\references{
http://en.wikipedia.org/wiki/Cholesky_decomposition
}

\keyword{linear model, multivariate normal, density, fast}