\name{HW-package}
\alias{HW-package}
\alias{HW}
\docType{package}
\title{Fastlm and dmvnorm
}
\description{
	The package contains 2 functions.One is for fast linear model regression and another for Multivariate normal distribution density.
}
\details{
\tabular{ll}{
Package: \tab HW\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-11-13\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{JerryPPA(Qiyuan Shi)
Who wrote it

Maintainer: Who to complain to <yourfault@somewhere.net>
~~ The author and/or maintainer of the package ~~
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
