% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forint.R
\name{get_bitcoin_price}
\alias{get_bitcoin_price}
\title{Accessing the Binance API to get the most recent price of a Bitcoin in USD.}
\usage{
get_bitcoin_price(retried = 0)
}
\description{
Accessing the Binance API to get the most recent price of a Bitcoin in USD.
}
