\name{Homework1-package}
\alias{Homework1-package}
\alias{Homework1}
\docType{package}
\title{
Homework 1
}
\description{
Advanced Statistics Computing Homework 1 which includes fastlm() and dmvnorm() two functions.
}
\details{
\tabular{ll}{
Package: \tab Homework1\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-11-08\cr
License: \tab GPL \cr
}
run fastlm() to fit a linear model and dmvnorm() to evaluate mvn density.
}
\author{
Yu Du

Maintainer: Yu Du<ydu@jhsph.edu>

}
\references{
Advanced Statistics Computing Class. Dr. Peng.
}

\keyword{ package }

